Fixed bug in LED_Static_Effects Bouncing Balls effect

+++ LED_Static_Effects.h
@@ -23,7 +23,7 @@ struct bouncing_mem_struct
   long*  ClockTimeSinceLastBounce = NULL;
   float* Dampening = NULL;
   bool*  ballBouncing = NULL;
-  int*   Height = NULL;
+  float*   Height = NULL;
 };
 
 class LEDStaticEffects
